pyproject.toml
[project]
name = "sideprojects-py"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "blackscholes>=0.2.0",
    "bokeh>=3.8.1",
    "ib-async>=2.1.0",
    "influxdb>=5.3.2",
    "ipython>=9.8.0",
    "kaleido>=1.2.0",
    "lightweight-charts>=2.1",
    "matplotlib>=3.10.8",
    "mplfinance>=0.12.10b0",
    "nbformat==5.10.4",
    "networkx>=3.6.1",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "plotly>=6.5.0",
    "pymysql>=1.1.2",
    "pyqt6==6.10.1",
    "pyqtgraph==0.14.0",
    "scikit-learn>=1.8.0",
    "scipy-stubs~=1.16.3",
    "seaborn>=0.13.2",
    "sns>=0.1",
    "sqlalchemy>=2.0.45",
]
